make sure that the last service ticket is actually valid
if remote cas server has been restarted, the service ticket will be
still in user session, but no longer valid. this will send the user
into an infinite redirect loop, because the session ticket will be
used instead of the new ticket in the url.

this makes sure that the invalid ticket does not remain in session
after first reuse attempt